+++ AllSuits.cs
@@ -4,11 +4,13 @@ using System.Text;
 using static suitsTerminal.StringStuff;
 using static suitsTerminal.Misc;
 using static suitsTerminal.ProcessRack;
+using static suitsTerminal.Bools;
 using System.Linq;
 using UnityEngine;
 using Component = UnityEngine.Component;
 using Steamworks.Ugc;
 using System.Collections;
+using TerminalApi;
 
 namespace suitsTerminal
 {
@@ -18,29 +20,82 @@ namespace suitsTerminal
         internal static List<UnlockableItem> UnlockableItems = new List<UnlockableItem>();
         internal static List<TerminalNode> suitPages = new List<TerminalNode>();
         internal static List<TerminalNode> otherNodes = new List<TerminalNode>();
+        internal static Dictionary<int,string> suitNameToID = new Dictionary<int,string>();
         internal static List<string> suitNames = new List<string>();
+        internal static List<string> favSuits = new List<string>();
         internal static List<Page> suitsPages = new List<Page>();
+        internal static bool favSuitsSet = false;
 
         internal static void GetList()
         {
+            suitNameToID.Clear();
             suitNames.Clear();
             weirdSuitNum = 0;
             foreach (UnlockableSuit item in allSuits)
             {
-                string SuitName = "";
+                string SuitName;
                 if (item.syncedSuitID.Value >= 0 && AddSuitToList(item))
                 {
                     SuitName = UnlockableItems[item.syncedSuitID.Value].unlockableName;
 
                     if(!SConfig.advancedTerminalMenu.Value)
                         SuitName = TerminalFriendlyString(SuitName);
-                    
-                    suitNames.Add(SuitName);
+
+                    suitNameToID.Add(item.syncedSuitID.Value, SuitName);
                 }
                 else
                 {
                     weirdSuitNum++;
                     suitsTerminal.X($"Skipping suit, either weird or locked.");
+                }
+            }
+
+            CheckForDuplicateSuitNames();
+            FixRack();
+            InitFavorites();
+            OldCommands.MakeRandomSuitCommand();
+        }
+
+        private static void CheckForDuplicateSuitNames()
+        {
+            List<string> fullList = new List<string>();
+            List<string> duplicatesFound = new List<string>();
+            // Populate suitNames based on suitNameToID
+            foreach (var kvp in suitNameToID)
+            {
+                // Check if the suit name already exists in suitNames
+                if (!fullList.Any(name => string.Equals(name, kvp.Value, StringComparison.OrdinalIgnoreCase)))
+                {
+                    // If not, add it directly to suitNames
+                    fullList.Add(kvp.Value);
+                }
+                else
+                {
+                    // If it does, append the integer after ^ to make it unique
+                    duplicatesFound.Add(kvp.Value);
+                }
+            }
+
+            CreateFinalSuitsList(duplicatesFound);
+        }
+
+        private static void CreateFinalSuitsList(List<string> duplicateSuitNames)
+        {
+            // Populate suitNames based on suitNameToID
+            foreach (var kvp in suitNameToID)
+            {
+                // If the suit name has not been encountered yet, add it to suitNames and the set
+                if (!duplicateSuitNames.Any(name => string.Equals(name, kvp.Value, StringComparison.OrdinalIgnoreCase)))
+                {
+                    suitNames.Add(kvp.Value);
+                }
+                else
+                {
+                    // If the suit name has been encountered, create a unique suit name with the associated key
+                    string uniqueSuitName = $"{kvp.Value}^({kvp.Key})";
+
+                    // Add both the original and unique suit names to suitNames
+                    suitNames.Add(uniqueSuitName);
                 }
             }
         }
@@ -50,9 +105,13 @@ namespace suitsTerminal
             if (!SConfig.enforcePaidSuits.Value)
                 return true;
 
-            if (!UnlockableItems[suit.syncedSuitID.Value].hasBeenUnlockedByPlayer && UnlockableItems[suit.syncedSuitID.Value].shopSelectionNode != null)
+            if (UnlockableItems[suit.syncedSuitID.Value].shopSelectionNode == null)
+                return true;
+
+            if (!UnlockableItems[suit.syncedSuitID.Value].spawnPrefab)
             {
-                suitsTerminal.X("Locked suit detected, not adding to list.");
+                suitsTerminal.X($"Locked suit [{UnlockableItems[suit.syncedSuitID.Value].unlockableName}] detected, not adding to list.");
+                suitsTerminal.X($"hasBeenUnlockedByPlayer: {UnlockableItems[suit.syncedSuitID.Value].hasBeenUnlockedByPlayer} \nalreadyUnlocked: {UnlockableItems[suit.syncedSuitID.Value].alreadyUnlocked}");
                 return false;
             }
             
@@ -94,8 +153,19 @@ namespace suitsTerminal
                     return false; // Keep the item
                 });
             }
+            if (SConfig.suitsSortingStyle.Value == "alphabetical")
+                OrderSuitsByName();
+            else if (SConfig.suitsSortingStyle.Value == "numerical")
+            {
+                OrderSuitsByID();
+            }
+            else if (SConfig.suitsSortingStyle.Value == "none")
+            {
+                suitsTerminal.Log.LogInfo("No sorting requested, host/client may be desynced.");
+            }
+            else
+                suitsTerminal.Log.LogError("Config failure, no sorting");
 
-            OrderSuitsByName();
         }
 
         private static void GetAllSuits()
@@ -117,12 +187,36 @@ namespace suitsTerminal
             allSuits = allSuits.OrderBy((UnlockableSuit suit) => UnlockableItems[suit.syncedSuitID.Value].unlockableName).ToList();
         }
 
+        private static void OrderSuitsByID()
+        {
+            allSuits = allSuits.OrderBy((UnlockableSuit suit) => suit.syncedSuitID.Value).ToList();
+        }
+
         internal static void InitSuitsListing()
         {
             GetAllSuits();
             GetList();
-            FixRack();
-            OldCommands.MakeRandomSuitCommand();
+        }
+
+        private static void InitFavorites()
+        {
+            List<string>favConfigList = SConfig.favoritesMenuList.Value.Split(',')
+                                      .Select(item => item.TrimStart())
+                                      .ToList();
+
+            favSuits.Clear();
+
+            foreach(string suitName in favConfigList)
+            {
+                if (suitNames.Contains(suitName) && !favSuits.Contains(suitName))
+                {
+                    favSuits.Add(suitName);
+                    suitsTerminal.X($"Added [{suitName}] to favorites list.");
+                }
+                else
+                    suitsTerminal.Log.LogWarning($"[{suitName}] not loaded to favorites. Suit is locked, already added, or not found.");
+                    
+            }
         }
 
         private static void HideBootsAndRack()
@@ -144,8 +238,7 @@ namespace suitsTerminal
                 GameObject clothingRack = GameObject.Find("Environment/HangarShip/NurbsPath.002");
                 GameObject.Destroy(clothingRack);
             }
-
-            
+   
         }
 
         private static void FixRack()
@@ -155,6 +248,10 @@ namespace suitsTerminal
             weirdSuitNum = 0;
             reorderSuits = 0;
             normSuit = 0;
+
+            List<string> customSuitNames = SConfig.suitsOnRackList.Value.Split(',')
+                                      .Select(item => item.TrimStart())
+                                      .ToList();
 
             HideBootsAndRack();
 
@@ -189,18 +286,18 @@ namespace suitsTerminal
                         }
                     }
 
-                    if (normSuit >= SConfig.suitsOnRack.Value && !rackSituated || (rackSituated && !isHanging) || SConfig.hideRack.Value)
+                    if (!ShouldShowSuit(item, customSuitNames))
                     {
                         ProcessHiddenSuit(component);
                         normSuit++;
                     }
-                    else if (rackSituated && isHanging)
+                    else if (rackSituated && ShouldShowSuit(item, customSuitNames))
                     {
                         ProcessHangingSuit(component);
                     }
-                    else if (normSuit < SConfig.suitsOnRack.Value && !rackSituated)
+                    else if (!rackSituated && ShouldShowSuit(item, customSuitNames))
                     {
-                        ProcessVisibleSuit(component, normSuit);
+                        ProcessVisibleSuit(component, showSuit);
                         normSuit++;
                     }
                     else
@@ -211,10 +308,8 @@ namespace suitsTerminal
                     if (showSuit == SConfig.suitsOnRack.Value && !rackSituated)
                     {
                         rackSituated = true;
-                        suitsTerminal.X($"Max suits are on the rack now. rack is situated \n^ ^\n .");
-                    }
-
-                    
+                        suitsTerminal.X($"Max suits are on the rack now. rack is situated, yippeee!!!");
+                    }   
                 }  
             }